splits component into their own files
